+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [STVP](https://mia.kiwi/projects/stvp).
 
-## [Unreleased]
+## [25.0.4] - 2025-08-26
 
 ### Added
 
@@ -20,3 +20,15 @@ and this project adheres to [STVP](https://mia.kiwi/projects/stvp).
 - APIv1 routes
 - Posts controller
 - Tags controller
+- RSS feed endpoint
+- Route to get metadata of all posts
+- Method `getPostsByMatchingPath` to `PostsContainerInterface` to find posts with similar paths, effectively allowing posts to be in hierarchical "folders"
+- `__toString` method to `Post` model
+- `getPostsByMetadata` method to `PostsContainerInterface` to get posts based on their metadata
+- `filterPosts` method to `PostsContainerInterface` to filter a set of posts based on a number of criteria
+- `PostsContainer` abstract class to hold common methods not reliant on the storage implementation
+- Unlisted posts
+
+### Changed
+
+- Moved `PostsContainer` methods of `FSPostsContainer`